requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
email-validator==2.2.0

# Vector DB (swap with real Qdrant in production)
# qdrant-client==1.11.0

# Embedding models (for production)
# sentence-transformers==3.1.1
# torch==2.4.1

# Advanced reranking (for production)
# FlagEmbedding==1.2.11

# Document parsing (for production)
# PyMuPDF==1.24.11
# pdfplumber==0.11.4
# python-docx==1.1.2

# LangChain integration (for production)
# langchain==0.3.0
# langchain-community==0.3.0
